Reorganise packages for schedule list
