+++ schedule/src/main/java/com/droidcon/schedule/ioc/ScheduleFragmentModule.kt
@@ -3,8 +3,8 @@ package com.droidcon.schedule.ioc
 import com.droidcon.commons.ioc.ChildFragmentScope
 import com.droidcon.commons.ioc.FragmentScope
 import com.droidcon.commons.conference.data.ioc.SessionizeModule
-import com.droidcon.schedule.ui.fragment.ScheduleDayFragment
-import com.droidcon.schedule.ui.fragment.ScheduleFragment
+import com.droidcon.schedule.ui.schedulelist.fragment.ScheduleDayFragment
+import com.droidcon.schedule.ui.schedulelist.fragment.ScheduleFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
