+++ schedule/src/main/java/com/droidcon/schedule/ioc/ScheduleModule.kt
@@ -1,9 +1,9 @@
 package com.droidcon.schedule.ioc
 
-import com.droidcon.schedule.ui.DayDividerItemDiffCallback
-import com.droidcon.schedule.ui.logic.GetScheduleTabs
-import com.droidcon.schedule.ui.SessionItemDiffCallback
-import com.droidcon.schedule.ui.SessionRowDiffItemCallback
+import com.droidcon.schedule.ui.schedulelist.recyclerview.DayDividerItemDiffCallback
+import com.droidcon.schedule.ui.schedulelist.logic.GetScheduleTabs
+import com.droidcon.schedule.ui.schedulelist.recyclerview.SessionItemDiffCallback
+import com.droidcon.schedule.ui.schedulelist.recyclerview.SessionRowDiffItemCallback
 import dagger.Module
 import dagger.Provides
