+++ schedule/src/main/java/com/droidcon/schedule/ioc/SearchSessionsFragmentModule.kt
@@ -1,7 +1,7 @@
 package com.droidcon.schedule.ioc
 
 import com.droidcon.commons.ioc.FragmentScope
-import com.droidcon.schedule.ui.fragment.SearchSessionsFragment
+import com.droidcon.schedule.ui.schedulelist.fragment.SearchSessionsFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
