+++ schedule/src/main/java/com/droidcon/schedule/ui/schedulelist/model/ScheduleDayModels.kt
@@ -1,4 +1,4 @@
-package com.droidcon.schedule.ui.model
+package com.droidcon.schedule.ui.schedulelist.model
 
 sealed class ScheduleState {
     data class Content(val sessionRows: List<SessionRow.Session>) : ScheduleState()
