+++ schedule/src/main/res/navigation/schedule_navigation.xml
@@ -5,7 +5,7 @@
         app:startDestination="@id/scheduleFragment">
     <fragment
             android:id="@+id/scheduleFragment"
-            android:name="com.droidcon.schedule.ui.fragment.ScheduleFragment"
+            android:name="com.droidcon.schedule.ui.schedulelist.fragment.ScheduleFragment"
             android:label="@string/schedule">
         <action
                 android:id="@+id/action_scheduleFragment_to_searchSessionsFragment"
@@ -13,6 +13,6 @@
     </fragment>
     <fragment
             android:id="@+id/searchSessionsFragment"
-            android:name="com.droidcon.schedule.ui.fragment.SearchSessionsFragment"
+            android:name="com.droidcon.schedule.ui.schedulelist.fragment.SearchSessionsFragment"
             android:label="SearchSpeakersFragment" />
 </navigation>
